# fix: in_span performance